Elasticsearch: reduce code duplication (#5137)
* Elasticsearch: reduce code duplication

* address comments

---------
